API Routes and stream depending on text file
